Rename `ok_if_let` to `if_let_some_result`
