+++ clippy_lints/src/if_let_some_result.rs
