+++ tests/ui/if_let_some_result.fixed
